Checkpoint runtime source and separate engine profiles
Version the current DotTalk++ runtime source graph in the development authority. Introduce independent NONE/LEGACY/LMDB index modes and LEAN/PROFESSIONAL/EDUCATIONAL/DEVELOPMENT products, invert xbase/xindex ownership, gate product command surfaces, and add tracked package allow-lists.

Include pydottalk/profile/LMDB proofs, installed-runtime DLL handling, AI Portal records, and project licensing status as To be determined. Verified NONE 7/7, LEGACY 5/5, LMDB 8/8, plus source policy.

+++ include/xindex/key_common.hpp
@@ -5,7 +5,7 @@
 namespace xindex {
 
 using Key   = std::vector<std::uint8_t>;
-using RecNo = std::uint32_t;
+using RecNo = std::uint64_t;
 
 struct KeyLess {
     bool operator()(const Key& a, const Key& b) const noexcept {
@@ -15,3 +15,6 @@ struct KeyLess {
 };
 
 } // namespace xindex
+
+
+
